Added RSS icon, updated podcast md, added new post
New post is not quiet ready, need to adjust image file directory but first I am seeing that file location works on the deployed website.

Added a RSS icon to home page so people can follow along with any changes. Also updated the podcast page's markdown because it was a bit of a mess, then added more content warnings to the podcast page which will forever be a work in progress.